clean up cruft from windows libdl implementation (#13)
we don't actually use that last_* stuff ... I think it may have been used in
the previous prototype of FFI::Platypus.

Second, we're just loading a library that is already loaded so that we can
find the address for the symbol.  Once that is done, we can free the
handle.  Better home that whoever attached the DLL doesn't free it of course.
If this becomes a problem we can perminently attach these DLLs.  I think
in practice what people are going to be using this feature for is to find
libc symbols, and libc shouldn't ever be detached.