move libtest -> t/ffi
This matches the convention used by the newer FFI::Build.  It makes sense
to put the testing library in the testing directory.  It also matches
the convention of putting the runtime ffi lib in ./ffi.  But the main
reason I am doing this is so that I can once again type l<TAB> instead of
lib/<TAB>.  Yeah.  I said it.