add 5.28 remove 5.10/5.12 from automated testing
yay travis finally got its act together and supports Perl 5.28
technically 5.8, 5.10 and 5.12 are supported (patches welcome!)
so long as the toolchain supports it.  automated testing on those
versions is starting to get harder (we already removed 5.8 for
that reason), and it takes a lot of time to run the regression
tests on so many different versions and configuration.  Also tbh
you should be upgrading to a more recent version of Perl.