combine fixed-string and record types
The fixed string was always represented and handled internally the
same as a record with no package (minor caveat: recently records were
made read-only as passed into a closure), so lets just make a fixed-string
a record internally, to simplify the internal code.

The only major change in behavior is to type_meta, and that has always
been "implementation dependent and subject to change".