fix bug where wrong IV size is used
The logic for the probe is correct, if we have 64 it IVs, we set
HAVE_IV_IS_64 to 1 and if not it is set to 0.  Unfortunately the
code checks if HAVE_IV_IS_64 is defined.

Fix the probe so that it does not set HAVE_IV_IS_64 at all for Perls
with 32 bit IVs.